The most significant changes in the code are the extensive modifications in the UserController.cs file and the adjustment of the bindingRedirect for the WebGrease library in the Web.config file.
In the UserController.cs file, several methods have been implemented to handle HTTP requests, including Index, Details, Create, Edit, and Delete. These methods are crucial for managing operations on user data.

In the Web.config file, the bindingRedirect for the WebGrease library has been modified. The oldVersion attribute has been changed to redirect any version of the WebGrease library from 0.0.0.0 to 1.6.5135.21930 to use version 1.6.5135.21930. This change ensures that the correct version of the library is used, which can prevent potential compatibility issues.

List of changes:

1. Addition of a reference to the Microsoft.CodeDom.Providers.DotNetCompilerPlatform library in the CRUD application 2.csproj file.
2. Removal of the reference to the Microsoft.CodeDom.Providers.DotNetCompilerPlatform library in the CRUD application 2.csproj file.
3. Removal of a comment at the end of the CRUD application 2.csproj file.
4. Implementation of various methods in the UserController class in the UserController.cs file.
5. Modification of the bindingRedirect for the WebGrease library in the Web.config file.